backend/requirements.txt
fastf1==3.3.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.1
ortools==9.10.4067
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
joblib==1.4.2
python-dotenv==1.0.1
requests==2.32.3
plotly==5.22.0
shap==0.45.1
supabase==2.28.3
websockets==13.1
